waste_prediction_module-v2/federated_client/requirements.txt
flwr
torch
numpy
python-dotenv
fastapi
uvicorn
